pyproject.toml
[tool.poetry]
name = "platform-proto"
version = "0.1.0"
description = "Hydra platform protobuf/gRPC definitions"
readme = "README.md"
authors = ["Hydra"]
license = "See LICENSE.txt"
packages = [{include = "platform_proto", from = "python"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.8"
protobuf = ">=4.21.0"
grpcio = ">=1.50.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = ">=1.50.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
